+++ source/blender/blenkernel/BKE_volume_render.h
@@ -27,6 +27,9 @@
  *  \ingroup bke
  *  \brief Volume datablock rendering and viewport drawing utilities.
  */
+
+#include "BLI_sys_types.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -38,16 +41,16 @@ struct VolumeGrid;
 
 bool BKE_volume_grid_dense_bounds(const struct Volume *volume,
                                   struct VolumeGrid *volume_grid,
-                                  ssize_t min[3],
-                                  ssize_t max[3]);
+                                  int64_t min[3],
+                                  int64_t max[3]);
 void BKE_volume_grid_dense_transform_matrix(const struct VolumeGrid *volume_grid,
-                                            const ssize_t min[3],
-                                            const ssize_t max[3],
+                                            const int64_t min[3],
+                                            const int64_t max[3],
                                             float mat[4][4]);
 void BKE_volume_grid_dense_voxels(const struct Volume *volume,
                                   struct VolumeGrid *volume_grid,
-                                  const ssize_t min[3],
-                                  const ssize_t max[3],
+                                  const int64_t min[3],
+                                  const int64_t max[3],
                                   float *voxels);
 
 /* Wireframe */
